ROX-14324: Use requester scope and target resource to validate cluster existence for NetworkGraph and NetworkPolicy
