+++ central/networkpolicies/service/service.go
@@ -15,6 +15,7 @@ import (
 	"github.com/stackrox/rox/central/networkpolicies/generator"
 	"github.com/stackrox/rox/central/networkpolicies/graph"
 	notifierDataStore "github.com/stackrox/rox/central/notifier/datastore"
+	"github.com/stackrox/rox/central/role/sachelper"
 	"github.com/stackrox/rox/central/sensor/service/connection"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/pkg/grpc"
@@ -58,6 +59,7 @@ func New(storage npDS.DataStore,
 		notifierStore:    notifierStore,
 		clusterStore:     clusterStore,
 		graphEvaluator:   graphEvaluator,
+		sacHelper:        sachelper.NewSacHelper(clusterStore, nil),
 		policyGenerator:  generator.New(storage, deployments, namespacesStore, globalFlowDataStore, networkTreeMgr, networkBaselines),
 	}
 }
