+++ central/role/sachelper/access.go
@@ -1,0 +1,8 @@
+package sachelper
+
+const (
+	hasFullAccess    = true
+	hasPartialAccess = false
+	fullAccess       = true
+	partialAccess    = false
+)
